+++ CHANGELOG.md
@@ -87,6 +87,7 @@ by a different Flowman project, you now need to explicitly import the project in
 ## Version 0.30.1 - 2023-03-23
 
 * github-379: [BUG] Parallel execution of multiple targets runs too many targets on Java 17
+* github-383: Flowman should preserve target ordering
 
 
 ## Version 0.30.0 - 2023-01-03